chore: expand dictionary helpers
