+++ core/dictionary.py
@@ -1,16 +1,15 @@
-"""Simple dictionary utilities for vocabulary categories."""
+"""Utility helpers for vocabulary category definitions."""
 
 import json
 from pathlib import Path
 from typing import Dict
 
-
-_DEF_PATH = Path("data/vocab_defs.json")
+_DEF_PATH = Path('data/vocab_defs.json')
 
 
 def load_definitions(path: Path = _DEF_PATH) -> Dict[str, str]:
     if path.exists():
-        with open(path, "r", encoding="utf-8") as f:
+        with open(path, 'r', encoding='utf-8') as f:
             return json.load(f)
     return {}
 
@@ -21,3 +20,15 @@ _DEFINITIONS = load_definitions()
 def describe(category: str) -> str | None:
     """Return a human readable description for a vocabulary category."""
     return _DEFINITIONS.get(category)
+
+
+def definitions() -> Dict[str, str]:
+    """Return a copy of all loaded definitions."""
+    return dict(_DEFINITIONS)
+
+
+def add_definition(category: str, text: str) -> None:
+    """Add or update a category definition and persist it."""
+    _DEFINITIONS[category] = text
+    with open(_DEF_PATH, 'w', encoding='utf-8') as f:
+        json.dump(_DEFINITIONS, f, indent=2)
