+++ core/prompt_engine.py
@@ -1,7 +1,5 @@
 import json
 import random
-import hashlib
-from datetime import datetime
 from pathlib import Path
 
 from .template_engine import TemplateRenderer, available_templates
@@ -69,43 +67,3 @@ class PromptEngine:
 def templates() -> list[str]:
     """Expose available templates."""
     return available_templates()
-
-
-def save_prompt_hash(prompt: str, log_path: Path = Path("logs/prompt_hashes.txt")) -> None:
-    hash_val = hashlib.sha256(prompt.encode()).hexdigest()
-    log_path.parent.mkdir(parents=True, exist_ok=True)
-    with open(log_path, "a", encoding="utf-8") as f:
-        f.write(f"{datetime.now().isoformat()} | {hash_val} | {prompt}\n")
-
-
-def save_prompts(prompts, format: str = "json", out_dir: str = "outputs", tag: str = "default", model: str | None = None) -> Path:
-    out_path = Path(out_dir)
-    out_path.mkdir(parents=True, exist_ok=True)
-    filename = f"batch_{datetime.now().strftime('%Y%m%d_%H%M%S')}_{tag}.{format}"
-    filepath = out_path / filename
-
-    if format == "json":
-        with open(filepath, "w", encoding="utf-8") as f:
-            json.dump(prompts, f, indent=2)
-    elif format == "txt":
-        with open(filepath, "w", encoding="utf-8") as f:
-            for p in prompts:
-                f.write(p + "\n")
-    elif format == "csv":
-        with open(filepath, "w", encoding="utf-8") as f:
-            f.write("prompt\n")
-            for p in prompts:
-                f.write(f"\"{p}\"\n")
-    elif format == "gpt":
-        payload = [
-            {"model": model or "gpt-4o", "messages": [{"role": "user", "content": p}]}
-            for p in prompts
-        ]
-        with open(filepath, "w", encoding="utf-8") as f:
-            json.dump(payload, f, indent=2)
-    else:
-        raise ValueError("Unsupported format")
-
-    for p in prompts:
-        save_prompt_hash(p)
-    return filepath
